requirements.txt
flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Bootstrap==3.3.7.1
